Add testlib method to use stuffers as io instead of sockets
Mutliple test classes repeat the same test setup code. Move
it to testlib so we only have to write it once.
